Stocks
The cost of stock on each day is given in an array A[] of size N. You can either buy or sell a stock on a particular day. You need to sell the previous stock before buying a new one. Find all the days on which you buy and sell the stock so that in between those days your profit is maximum.

There are no two days with same price of the stock.

Input:
7
100 180 260 310 40 535 695

Output:
0 3
4 6